Restore auth and web-assets tests
