+++ crates/api-ui/src/tests/mod.rs
@@ -1,3 +1,4 @@
+pub mod auth;
 pub mod common;
 pub mod dashboard;
 pub mod databases;
