--- crates/history/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "embucket_history"
-version = "0.1.0"
-edition = "2021"
-license-file = { workspace = true }
-
-[dependencies]
-embucket_utils = { path = "../utils" }
-bytes = { workspace = true }
-chrono = { workspace = true, features = ["serde"] }
-serde = { workspace = true }
-serde_json = { workspace = true }
-utoipa = { workspace = true }
-uuid = { workspace = true }
-snafu = { workspace = true }
-tracing = { workspace = true }
-slatedb = { workspace = true }
-async-trait = { workspace = true }
-futures = { workspace = true }
-
-[dev-dependencies]
-tokio = { workspace = true }
-object_store = { workspace = true }
-mockall = "0.13.1"
-
-[lints]
-workspace = true
